Refitting rework
Now you pay exactly the difference between the module you want replaced and the module you're adding. This also means that some refits can be free if the module you currently have is more expensive than the one you're replacing it with. This applies to every single module, including engines and armour tech, so refitting them should be common now.